requirements.txt
Flask==2.3.2
PyPDF2
python-docx
Pillow